inputs taken

+++ lib/models/user.dart
@@ -2,7 +2,7 @@ import 'package:thrifty/models/expensemodel.dart';
 
 class User {
   final String uid;
-  int maxAmount;
+  double maxAmount;
   List<Expense> expenseList;
   User({this.uid});
 }
